秒杀业务-上架秒杀商品Done

+++ gulimall-coupon/src/main/java/com/lif314/gulimall/coupon/entity/SeckillSkuRelationEntity.java
@@ -10,7 +10,7 @@ import lombok.Data;
 
 /**
  * 秒杀活动商品关联
- * 
+ *
  * @author lif314
  * @email lifer314@163.com
  * @date 2022-02-07 22:09:08
@@ -44,11 +44,11 @@ public class SeckillSkuRelationEntity implements Serializable {
 	/**
 	 * 秒杀总量
 	 */
-	private BigDecimal seckillCount;
+	private Integer seckillCount;
 	/**
 	 * 每人限购数量
 	 */
-	private BigDecimal seckillLimit;
+	private Integer seckillLimit;
 	/**
 	 * 排序
 	 */
