+++ gulimall-seckill/src/main/java/com/lif314/gulimall/seckill/vo/SeckillSkuRelationVo.java
@@ -28,11 +28,11 @@ public class SeckillSkuRelationVo {
     /**
      * 秒杀总量
      */
-    private BigDecimal seckillCount;
+    private Integer seckillCount;
     /**
      * 每人限购数量
      */
-    private BigDecimal seckillLimit;
+    private Integer seckillLimit;
     /**
      * 排序
      */
